Update Chart plugin version and improve code readability
Updated the Chart plugin to version 5.0 and modified the minimum and maximum version of Ilias. In addition, multiple adjustments were made to improve the legibility of the code, such as adjusting indentation for consistency, adding a couple of useful scripts to the html template, and modifying the constants to be public for better accessibility.