Support ILIAS 9 (#14)
* Update Chart plugin version and improve code readability

Updated the Chart plugin to version 5.0 and modified the minimum and maximum version of Ilias. In addition, multiple adjustments were made to improve the legibility of the code, such as adjusting indentation for consistency, adding a couple of useful scripts to the html template, and modifying the constants to be public for better accessibility.

* Updated script.js to handle empty chart_divs array

This commit modifies the script.js file to include a check for whether the chart_divs array is empty before assigning the first item's id to 'min'. This ensures that an error isn't thrown when trying to access an index that does not exist in the array.

* Update requirements and changelog

The plugin requirements have been updated to reflect new compatibility with ILIAS 9 and PHP 8.1. A new entry in the changelog has been added to indicate that the code has been refactored for ILIAS 9 readiness.

* Update plugin compatibility and clean up code

The version compatibility of the Chart plugin for ILIAS has been updated to now include ILIAS 8 in addition to ILIAS 9. The PHP version required for ILIAS 8 is also specified. Additionally, some code in class.ilChartPluginGUI.php has been refactored to be cleaner and easier to read. A method from class.ilChartPlugin.php that was returning Javascript files has also been removed.

* Change access modifiers for constants in ChartPluginGUI

Updated class constants in class.ilChartPluginGUI.php to be private in order to encapsulate variables within the class and improve security.